// union-talk-frontend/src/services/generated/file/models/uploadSessionRespVO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * 上传任务响应参数
 */
export interface UploadSessionRespVO {
  /** 上传任务id */
  sessionId?: string
  /** 上传token */
  uploadToken?: string
  /** 会话id */
  conversationId?: string
  /** 文件夹id */
  folderId?: string
  /** 资产文件id */
  assetId?: string
  /** 文件名称 */
  fileName?: string
  /** 文件扩展名 */
  fileExt?: string
  /** MIME类型 */
  mimeType?: string
  /** 文件大小 */
  fileSize?: string
  /** 分片大小 */
  chunkSize?: string
  /** 分片数量 */
  chunkCount?: number
  /** 已上传分片数量 */
  uploadedChunkCount?: number
  /** 已上传分片序号列表 */
  chunkIndexList?: number[]
  /** 上传状态 */
  status?: string
  /** 创建时间 */
  createdAt?: string
  /** 更新时间 */
  updatedAt?: string
}